+++ apps/api/src/features/assistant-tools/tools.module.ts
@@ -1,12 +1,12 @@
 import {getMarketIndicatorTool} from "./tools/get-market-indicator.tool";
 import {addKnowledgeTool} from "./tools/add-knowledge.tool";
 import {getKnowledgeTool} from "./tools/get-knowledge.tool";
-import {createReminderTool} from "./tools/notifications-tools/create-reminder.tool";
+import {createReminderTool} from "./tools/create-reminder.tool";
 
 export enum AssistantTool {
     MARKET_INDICATORS = 'MARKET_INDICATORS',
     ADD_KNOWLEDGE = 'ADD_KNOWLEDGE',
-    FIND_RELEVANT_KNOWLEDGE = 'FIND_RELEVANT_KNOWLEDGE',,
+    FIND_RELEVANT_KNOWLEDGE = 'FIND_RELEVANT_KNOWLEDGE',
     CREATE_REMINDER = 'CREATE_REMINDER'
 }
